ab 8: dded wner + eature + aptruct mapping + relationships

+++ src/main/java/com/example/lab7/service/CarService.java
@@ -2,64 +2,53 @@ package com.example.lab7.service;
 
 import com.example.lab7.dto.CarDto;
 import com.example.lab7.entity.Car;
+import com.example.lab7.mapper.CarMapper;
 import com.example.lab7.repository.CarRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class CarService {
 
     private final CarRepository carRepository;
-
-
-    public CarDto toDto(Car car) {
-        return CarDto.builder()
-                .id(car.getId())
-                .brand(car.getBrand())
-                .model(car.getModel())
-                .year(car.getYear())
-                .price(car.getPrice())
-                .build();
-    }
-
-    public Car toEntity(CarDto dto) {
-        return Car.builder()
-                .id(dto.getId())
-                .brand(dto.getBrand())
-                .model(dto.getModel())
-                .year(dto.getYear())
-                .price(dto.getPrice())
-                .build();
-    }
-
+    private final CarMapper carMapper;
 
     public List<CarDto> getAll() {
-        return carRepository.findAll().stream().map(this::toDto).collect(Collectors.toList());
+        return carMapper.toDtoList(carRepository.findAll());
     }
 
     public CarDto getById(Long id) {
-        return carRepository.findById(id).map(this::toDto).orElse(null);
+        return carRepository.findById(id)
+                .map(carMapper::toDto)
+                .orElse(null);
     }
 
     public CarDto create(CarDto dto) {
-        Car car = toEntity(dto);
-        return toDto(carRepository.save(car));
+        Car car = carMapper.toEntity(dto);
+        Car saved = carRepository.save(car);
+        return carMapper.toDto(saved);
     }
 
     public CarDto update(Long id, CarDto dto) {
-        Car existing = carRepository.findById(id).orElseThrow();
-        existing.setBrand(dto.getBrand());
-        existing.setModel(dto.getModel());
-        existing.setYear(dto.getYear());
-        existing.setPrice(dto.getPrice());
-        return toDto(carRepository.save(existing));
+        if (!carRepository.existsById(id)) {
+            return null;
+        }
+
+        Car car = carMapper.toEntity(dto);
+        car.setId(id); // важно!
+
+        Car updated = carRepository.save(car);
+        return carMapper.toDto(updated);
     }
 
-    public void delete(Long id) {
+    public boolean delete(Long id) {
+        if (!carRepository.existsById(id)) {
+            return false;
+        }
         carRepository.deleteById(id);
+        return true;
     }
 }
